Cleanup and separation into modules
